requirements.txt
flask==2.3.3
flask-cors==4.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
urllib3==2.0.7
dnspython==2.4.2
cryptography==41.0.7
python-whois==0.8.0
pillow==10.1.0
validators==0.22.0
google-api-python-client==2.108.0
google-auth==2.23.4
aiohttp==3.9.1
asyncio-throttle==1.0.2
flask-socketio==5.3.6
selenium==4.15.2
webdriver-manager==4.0.1
